Create a separate test file for incremental error handling
